Blockchain/wasteland/backend/requirements.txt
fastapi
pymysql
python-dotenv
python-multipart
uvicorn
sqlalchemy
strawberry-graphql